pipeline/requirements.txt
argparse==1.4.0
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
python-dotenv==0.21.0
pytest>=6.0.0
PyYAML==6.0
requests==2.25.1
scikit-learn==1.2.2
seaborn==0.12.2
typing==3.7.4.3
s3fs==2023.1.0
boto3>=1.24.0,<1.25.0
aiobotocore==2.4.2
imblearn
xgboost==2.0.3
